# sensor_hub/api/openapi.yaml
openapi: 3.1.0
info:
  title: Sensor Hub API
  version: 1.0.0
  description: >-
    API for Sensor Hub. Provides endpoints for managing sensors,
    querying readings with automatic aggregation, updating application
    properties, managing users and RBAC, dashboards and more. Authentication is
    cookie-based with CSRF protection, as well as API Key for service-to-service
    communication. WebSocket endpoints are available for real-time updates of
    sensor readings and properties changes.
servers:
  - url: http://localhost:8080/api
security:
  - cookieAuth: []
    csrfToken: []
  - apiKeyAuth: []
paths:
...
  /email/smtp:
    get:
      tags:
        - email
      summary: Get the SMTP settings
      description: >-
        Returns the SMTP server alert and notification emails are sent
        through, whether a password is stored, and how the last send went.
        The password itself is never returned.
      operationId: getEmailSettings
      x-required-permission: manage_email
      responses:
        '200':
          description: The SMTP settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailSettings'
        '401':
          description: Not authenticated
        '403':
          description: Insufficient permissions
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - email
      summary: Update the SMTP settings
      description: >-
        Replaces the SMTP settings. The password is held encrypted in the
        secret store: a non-empty value other than "****" sets it, an empty
        string clears it, and omitting it or sending "****" leaves it
        unchanged.
      operationId: updateEmailSettings
      x-required-permission: manage_email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailSettings'
      responses:
        '200':
          description: The updated SMTP settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailSettings'
        '400':
          description: >-
            Invalid settings: an empty host, a port outside 1 to 65535, an
            unknown security mode or a from address that is not an email
            address.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
        '403':
          description: Insufficient permissions
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /email/smtp/test:
    post:
      tags:
        - email
      summary: Send a test email
      description: >-
        Sends a message titled "Sensor Hub test email" to the caller's own
        email address through the stored SMTP settings, and records the
        outcome as a send.
      operationId: sendTestEmail
      x-required-permission: manage_email
      responses:
        '200':
          description: The SMTP server accepted the test email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '400':
          description: >-
            The caller has no email address, or the SMTP settings have no host
            or no usable password, so nothing was sent.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
        '403':
          description: Insufficient permissions
        '502':
          description: The send failed; `message` holds the SMTP server's error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Human-readable error message.
      required:
        - message
...
    EmailSettings:
      type: object
      description: >-
        The SMTP server alert and notification emails are sent through. The
        hub sends nothing until a host is set and a password is stored.
      properties:
        host:
          type: string
          description: SMTP server host name.
          example: "email-smtp.eu-west-1.amazonaws.com"
        port:
          type: integer
          minimum: 1
          maximum: 65535
          description: SMTP server port, usually 587 for starttls and 465 for implicit_tls.
          example: 587
        security:
          type: string
          enum: [starttls, implicit_tls, none]
          default: starttls
          description: >-
            How the connection is protected. starttls upgrades a plain
            connection with STARTTLS and fails if the server does not offer
            it; implicit_tls opens TLS first. Both verify the server's
            certificate against the system's trusted roots and that it was
            issued for the host. none sends everything, the password
            included, unencrypted.
        username:
          type: string
          description: >-
            Login name for SMTP authentication. When empty the hub does not
            authenticate.
        from_address:
          type: string
          description: >-
            Address the emails are sent from, such as alerts@example.com.
            Empty until it is set.
          example: "alerts@example.com"
        password:
          type: string
          writeOnly: true
          description: >-
            SMTP password, write-only: no response carries it. It is held
            encrypted in the secret store. A non-empty value other than "****"
            sets it, an empty string clears it, and omitting it or sending
            "****" leaves the stored password unchanged.
        password_status:
          type: string
          readOnly: true
          enum: [unset, set, needs_reentry]
          description: >-
            Whether a password is stored. needs_reentry means the stored
            password does not decrypt under the hub's current key and has to
            be entered again; until then nothing is sent.
        last_error:
          type: string
          readOnly: true
          nullable: true
          description: The error the last send failed with, or null when the last send succeeded or none was tried.
        last_sent_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
          description: When an email was last sent successfully, or null if none has been.
      required:
        - host
        - port
        - security
        - username
        - from_address

    # =========================================================================
    # Dashboard Schemas
    # =========================================================================
...
    SuccessMessage:
      type: object
      description: Generic success response
      properties:
        message:
          type: string
      required:
        - message
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sensor_hub_session
      description: Session cookie set after successful login. Configurable via AuthSessionCookieName.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: CSRF token returned from login, required for state-changing operations (POST, PUT, PATCH, DELETE).
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >-
        API key for programmatic and CLI access.
        Exempt from CSRF validation. Create keys via the
        API Keys management endpoints.

  # Optional: group endpoints for UIs and tooling
...